+++ CHANGELOG.md
@@ -2,11 +2,11 @@
 
 Change Logs.
 
-## **
+## 1.4.9 beta
 
 * Can see selected language and hashtags easily in watermark text field.
 
-## 1.4.8 beta
+## 1.4.8
 
 * Count the number of tweet text count more accurately.
 * Can open link directly from timeline view.